Extract shared boilerplate from sync cursors into WithFetch mixin
Add WithFetch class to result_set.py as the sync counterpart to
WithAsyncFetch, consolidating ~500 lines of duplicated properties,
lifecycle methods, and fetch methods from 5 sync cursor types.

Closes #669
